Remove health check API and related documentation
