+++ docs/api/openapi.yaml
@@ -9,26 +9,6 @@ servers:
   - url: http://localhost:8080
 
 paths:
-  /api/health:
-    get:
-      summary: Health check
-      operationId: getHealth
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                required: [status, message]
-                properties:
-                  status:
-                    type: string
-                    example: ok
-                  message:
-                    type: string
-                    example: Server is running
-
   /api/files/{name}:
     put:
       summary: Upload (store) a file by name
