--- internal/adapter/http/health_handler.go
@@ -1,6 +1,0 @@
-package httpadapter
-
-// TODO:
-// - Implement HTTP handler for GET /api/health.
-// - Call usecase and return JSON {status,message}.
-
