+++ internal/adapter/http/router.go
@@ -2,7 +2,6 @@ package httpadapter
 
 // TODO:
 // - Wire routes using net/http ServeMux:
-//   - GET  /api/health
 //   - PUT  /api/files/{name}
 //   - GET  /api/files/{name}
 // - Keep routing and HTTP concerns in this adapter.
