--- internal/infra/storage/fs/filestore_test.go
@@ -1,5 +1,0 @@
-package fs
-
-// TODO:
-// - Add tests using t.TempDir() for Put/Get/Exists.
-
