--- internal/usecase/health_check.go
@@ -1,6 +1,0 @@
-package usecase
-
-// TODO:
-// - Implement HealthCheck usecase (returns status/message).
-// - Optionally include dependency checks (e.g., storage) later.
-
